Working on mutator Invoice

+++ database/migrations/2022_10_26_054331_create_invoices_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             // $table->foreignId('teacher_classroom_id')->nullable()->constrained()->nullOnDelete();
             $table->foreignId('school_year_id')->nullable()->constrained()->nullOnDelete();
             $table->foreignId('classroom_id')->nullable()->constrained()->nullOnDelete();
-            $table->tinyInteger('payment_for_month');
+            $table->Integer('payment_for_month');
             $table->Integer('amount');
             $table->Integer('fine_amount');
             $table->Integer('personal_discount');
